feat(pkictl.eecert): add dynamic naming for end entity certs
Allow user to specify a custom name for the results of
`pkictl eecert` commands. The name that is chosen upon request, is the
name that sticks for the remainder of the commands. This allows multiple
end entity cert requests from the same configuration files to have
different and easy to identity names.

fixes #7