fix(pkictl.eecert.import): chains moved outside of certs folder
certs folder is now reserved for single cert-per-file files and includes
the entire chain from root to end-entity. This is because all those
files need hashing eventually and separating by directory makes that
easier and reduces redundancies. Intermediate and CA cert chains are for
specific program usage and thus OK to leave separated. They don't need
to be hashed.