fix(pkictl.eecert.import): dynamic naming, don't overwrite files
Import policy is non-clobbering by default. If an incoming file has the
same name as an already existing file, it's name is appended with a
counter. It is then checked against all existing files with that same
base name and if the new file is a duplicate, it is deleted keeping the
original intact with it's old name in case a program is already using
it. This works with keys, certificates, chains, everything.

While most files should in practice should show up with different file
names if they are indeed for different purposes, because parts of the
naming scheme are extracted from the certificate's subject information,
duplicate names are possible. There are two common scenarios where
handling this is useful:

* Replacing a obsolete/revoked key
    * Good practice is to not delete old keys, but to let CRL's
      handle their status. Counter suffixes easily show the keys
      history on this machine.
* Keys for the same host and purpose, for different but similar programs
    * This could result in what appears to the the exact same subject
      information for both keys, differing only in the key information,
      but just for different program usage.

fixes #15