fix(pkictl,pkictl_test): keep PKCS12 import directory pure
Don't modify or make assumptions as to the sub-folders this dir has
(i.e. ./export). The setting for PKICTL_IMPORT_DIR must go all the way
to the files themselves. This includes appending ./export to the CA
directory if leaving in the default location that `pkictl genpkcs12`
puts them in initially.