feat(pkictl.eecert.import): strip bag/cert headers from certs/keys
This information is left alone in the certs/keys maintained by the CA.
Openssl can handle files with that info just fine and it's kept in to
make it very human readable. It is stripped however for all
certs/keys/chains being imported via a PKCS#12 file by default because
many programs can get confused by that info.