fix(pkictl.eecert.import): don't duplicate intermediates cert files
In addition to checking against files of the same name as the incoming
import, we now also have a special additional check of intermediate cert
files to check for duplicates accross different file names. Different
certs coming from the same CA will produce files with different names
but same contents.

Default behavior is same as other de-dupe functions: to keep the
original existing file intact first if the new file is the same;
better compatibility in case a program is already are using it.

fixes #26